feat(client): game card with squiz v1 style
#2
