+++ packages/client/package.json
@@ -20,6 +20,7 @@
     "@react-navigation/stack": "^5.9.0",
     "expo": "^39.0.0",
     "expo-av": "~8.6.0",
+    "expo-linear-gradient": "~8.3.0",
     "expo-status-bar": "~1.0.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
