+++ packages/client/src/components/Card/index.ts
@@ -1,0 +1,3 @@
+import GameCard from './GameCard';
+
+export { GameCard };
