+++ packages/client/src/screens/Home.tsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from 'react';
 import { View, StyleSheet, ActivityIndicator } from 'react-native';
 import io from 'socket.io-client';
 
-import Card from '../components/Card';
+import { GameCard } from '../components/Card';
 import CenterContainer from '../components/CenterContainer';
 import Text from '../components/Text';
 import getEnv from '../constant/index';
@@ -54,7 +54,7 @@ export default function Home({ navigation }: Props) {
       </Text>
       <View style={styles.cardsContainter}>
         {rooms.map((room) => (
-          <Card
+          <GameCard
             key={room.id}
             onPress={() => navigate(room.id)}
             players={room.players}
